chore: move VerifiedExecutableAttestedTransaction to iota-core
