+++ crates/iota-core/src/authority/authority_per_epoch_store.rs
@@ -42,8 +42,7 @@ use iota_types::{
     effects::TransactionEffects,
     error::{IotaError, IotaResult},
     executable_transaction::{
-        CertificateProof, ExecutableTransaction, VerifiedExecutableAttestedTransaction,
-        VerifiedExecutableTransaction,
+        CertificateProof, ExecutableTransaction, VerifiedExecutableTransaction,
     },
     global_state_hash::GlobalStateHash,
     iota_system_state::epoch_start_iota_system_state::{
@@ -126,6 +125,7 @@ use crate::{
     post_consensus_validation,
     signature_verifier::*,
     stake_aggregator::StakeAggregator,
+    transaction_manager::VerifiedExecutableAttestedTransaction,
 };
 
 /// The key where the latest consensus index is stored in the database.
