+++ crates/iota-types/src/executable_transaction.rs
@@ -2,12 +2,9 @@
 // Modifications Copyright (c) 2024 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-use std::ops::Deref;
-
 use serde::{Deserialize, Serialize};
 
 use crate::{
-    attestation::Attestation,
     committee::EpochId,
     crypto::AuthorityStrongQuorumSignInfo,
     message_envelope::{Envelope, TrustedEnvelope, VerifiedEnvelope},
@@ -81,58 +78,5 @@ impl VerifiedExecutableTransaction {
 
     pub fn gas_budget(&self) -> u64 {
         self.data().transaction_data().gas_budget()
-    }
-}
-
-/// Wraps a [`VerifiedExecutableTransaction`] with its pre-consensus
-/// [`Attestation`] (if any). Carrying the full attestation lets downstream code
-/// consult both scheduling metadata and attestor identity / observed object
-/// versions.
-#[derive(Clone, Debug)]
-pub struct VerifiedExecutableAttestedTransaction {
-    tx: VerifiedExecutableTransaction,
-    /// `None` for unattested transactions (e.g., `UserTransactionV1`).
-    attestation: Option<Attestation>,
-}
-
-impl VerifiedExecutableAttestedTransaction {
-    pub fn new(tx: VerifiedExecutableTransaction, attestation: Option<Attestation>) -> Self {
-        Self { tx, attestation }
-    }
-
-    /// Returns the attached attestation, or `None` if the transaction was
-    /// not attested.
-    pub fn attestation(&self) -> Option<&Attestation> {
-        self.attestation.as_ref()
-    }
-
-    /// Returns the attestor's estimated computation cost, or `None` if the
-    /// transaction was not attested.
-    pub fn attested_computation_cost(&self) -> Option<u64> {
-        self.attestation
-            .as_ref()
-            .map(|a| a.estimated_computation_cost())
-    }
-
-    /// Consume the wrapper and return its parts.
-    pub fn into_parts(self) -> (VerifiedExecutableTransaction, Option<Attestation>) {
-        (self.tx, self.attestation)
-    }
-}
-
-impl From<VerifiedExecutableTransaction> for VerifiedExecutableAttestedTransaction {
-    fn from(tx: VerifiedExecutableTransaction) -> Self {
-        Self {
-            tx,
-            attestation: None,
-        }
-    }
-}
-
-impl Deref for VerifiedExecutableAttestedTransaction {
-    type Target = VerifiedExecutableTransaction;
-
-    fn deref(&self) -> &Self::Target {
-        &self.tx
     }
 }
